Split update and commit steps
This creates two separate commands - one update step which simply
modifies the source files with the new version, and an additional commit
step which will commit all modified files to the repository. This is
done in order to acheive continuous integration builds and nightly
builds without having to create a new commit everytime.

Also, the update step takes an optional format argument so the user can
specify the version format to be saved in the source files. This is
useful in cases where the build framework expects the version string to
be in a certain format.