First commit to move instance attributes to class
This mainly addresses the compare_build attribute, so you can set it at
the class level in your application and decide at one level whether to
use the comparator or not.

Similarly, this adds a new enable_prerelease_parser to decide whether to
use the prerelease parser in your application or disable it to use your
own custom strings. Support for these class attributes is not yet in
place, since this is only an initial commit.

Right now, you can selectively enable or disable the compare_build
option, but it's at an instance level rather than at the class level.